Create the initial API.
Create the initial API @tgross and I agreed on, with some minor tweaks.
See the original:
https://gist.github.com/paddyforan/bf1520bfcd19c5c9ac81

The tweaks:

* Don't bother supporting the top level logger. It doesn't make sense
to, with Sentry configuration as part of the game. People can just
create their own global Logger instance. Whatever.
* No need for a helper subpackage, we can just expose a call to modify
the calldepth being used. Presumably the same calldepth will be used for
the package, even if there are a couple edge-cases.

The output should remain the same, and the concerns we talked about are
all addressed.

I have yet to figure out how to call through to Sentry in a clean way
while still providing data. Any thoughts would be appreciated.