Add Sentry support.
Create a helper method (asSentryLevel) that will turn a Level into its
corresponding raven.Severity.

Add tags and meta properties to the Logger type to support per-call
tagging and extra data for Sentry.

Add a packagePrefixes property to allow callers to customise what Sentry
designates as "in-app" or not.

Create a makeCopy helper that will create a copy of a Logger, for use in
setting per-call tags and metadata. We need to copy it because maps and
slices are both reference types, which means that calls would
accidentally bleed into each other, which is not what we want.

Create an AddTags method on Logger that returns a copied Logger.
Basically, this will let us do stuff like
logger.AddTags(map[string]string{"mytag": "value"}).Error("Something
happened") and have it only apply to that call.

Create an AddMeta method on Logger that returns a copied Logger. This
can be used to pass additional data into Sentry that it wouldn't make
sense to surface to our log files. We can do stuff like
logger.AddMeta(raven.NewHttp(r)).Errorf("Error serving request: %+v\n",
err) to include the request that caused the error in the Sentry message.

Create a SetPackagePrefixes method on Logger that will set the package
prefixes that Sentry should consider "in-app". I'm not 100% sure what
this does, but it was a surfaced option in the sentry lib, and seemed
important, so I exposed it.

Create a helper method that will send output to Sentry. This
automatically composes a Sentry message, takes a Stacktrace (because if
it's getting sent to Sentry, it should be a warning or error, anyways,
and we want stacktraces on those), loops through the arguments passed
(if it came from an Errorf or Warnf call) and pulls out any it can turn
into Sentry metadata, so we can naturally annotate the error. It also
adds any metadata that's associated with the logger (from AddMeta). Then
it sets the tags for the call, sets the severity from the Logger Level,
and finally sends the log to Sentry.

Create a helper method that will turn an interface{} into a
raven.Interface type, if it can. Right now, it supports errors (which
generate a stacktrace) and http.Request pointers.

Call our toSentry helper from the appropriate methods.