Fix sentry.
Fix our sentry setup code to skip the verification of TLS, because our
sentry cert isn't recognised by the OS for whatever reason.

Add a flcok property to our logger that is used as a file lock, so
writes to the file don't step on each other.

Add a SetRelease method on the Logger that allows us to set the Release
that will be sent to Sentry, so we know which versions have bugs in
them.

Fix our log/logf functions to pass in the level being logged, instead of
using the minimum level to log.

Use fmt.Sprintln instead of fmt.Sprint, so spaces get included as
expected.

Fix our calldepths to be accurate.

Fix our call to formatHeader to use the passed in level, instead of the
log-configured level, so it prints out the right level for the log
message.

Add more context to the stacktraces.

Fix Sentry formatting so things appear as expected in Sentry.

Check for errors from Sentry and log them when recording a Sentry event
fails.