Tests, minor tweaks, and fixes.
Add test coverage to make sure that we're getting the right output when
we're supposed to get output, and nothing when we're not supposed to get
output.

Update the formatHeader method to be a function, as we weren't using the
Logger anyways, and I didn't want to set up testing for something we
weren't using. Minimal design through laziness!

The biggest change here is that I stopped using pointers to our Logger
type, and instead just resolved that everything would be copied a lot.
Our main use case (things are set up at the beginning and never modified
during runtime) works well with this approach, and we get a lot less
lock-thrashing. It also makes things like updating the levels more
intuitive (that bit me while testing, previously) because our Level type
is just an alias of the string type, meaning it's immutable, so we can't
just use a pointer to it, update it, and expect it to update
everywhere--we're overwriting the pointer, not the data found at the
pointer.