Add net/context helper.
We want to be able to easily grab a logger from the
golang.org/x/net/context#Context type we're passing around. Rather than
implementing it in each and every application, I thought it made sense
to have a helper in the logging package itself that had a sane default
behaviour.