Update tests to work with 1.5's vendoring strategy.
Go 1.5's vendoring strategy puts everything in the /vendor/ directory
and treats that as a separate GOPATH, so the import strings stay the
same. However, our tests naively made assumptions about the relationship
of the GOPATH and the file that broke when the package became vendored.

To remedy this, I updated the tests to be less brittle and more aware of
their location in your filesystem, instead of just assuming it.