Change all our Set* methods to With* methods.
It came up that our Set* methods were misleading (#7) may be misleading
or confusing, and that it would be better to have them as With* methods.

I updated the names, updated the godoc comments, and made sure the tests
still pass (that line number thing is annoying, but what can you do?).
This should make things clearer.