Use the upstream raven-go, remove vendors.
We've learned it's a mistake to vendor in libraries, and we should
instead only do it in binaries. So drop the vendor directory.

We've also run into the issue that getsentry/raven-go was forked to
DramaFever/raven-go as a soft fork, but the example subpackage imports
getsentry/raven-go. That means when vendoring, we wind up with _both_
getsentry/raven-go and DramaFever/raven-go, which is suboptimal.

The reason we have the fork is for a DramaFever-specific dependency
management strategy that some of the projects use, that involves
downloading and manually placing code in the GOPATH. We want to avoid
turning the fork into a hard fork, and we want to avoid vendoring two
copies of the same code. So what we're going to do is import the
upstream version; when vendoring, it doesn't matter if they delete the
repo, we still have a copy. When using the other strategy, the
DramaFever fork will be downloaded, but it'll be placed in the upstream
path in the GOPATH. That means this code should work with either
strategy, once we point to the upstream path.